+++ CHANGELOG.md
@@ -32,6 +32,20 @@ All notable changes to this project will be documented in this file.
 
 - "Universal" node: group.play.snap now expects snap in payload and not in msg.snap!
 
+## [3.1.5] 2020-05-27
+
+### Changed
+
+- "My Sonos" node: bugfix get.items for Apple playlists
+
+## [3.1.4] 2020-05-19
+
+### Added
+
+- "My Sonos" node: get.items added radioId for TuneIn stations
+
+- "Universal node": for play.notification changed order volume/play
+
 ## [3.1.3] 2020-05-13
 
 ### Added